requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
sentence-transformers
rank_bm25
python-dotenv
streamlit
ragas
openai
pyyaml
